Move the custom gui module to the root of the project
